Fix docs generation

+++ docs.cjs
@@ -20,6 +20,7 @@ const fs = require('fs-extra');
 const simpleGit = require('simple-git');
 const git = simpleGit();
 const shell = require('shelljs');
+const { type } = require('os');
 
 /**
  * Exec a shell command
@@ -33,6 +34,7 @@ function exec(command) {
       if (code !== 0) {
         console.error(stderr);
       }
+      console.info(stdout);
       resolve(code);
     });
   })
@@ -54,21 +56,24 @@ const config = {
       ignoreFiles: ['README.md', '_sidebar.md'],
       typedocSourceDir: 'typedocs',
       typedocTargetDir: 'lightning-core',
-      typedocBuildRequired: true
+      typedocBuildRequired: true,
+      docType: 'L2'
     },
     {
       gitURL: 'https://github.com/rdkcentral/Lightning-UI',
       gitBranch: 'main',
       sourceDir: 'docs',
       targetDir: 'lightning-ui-reference',
-      ignoreFiles: ['README.md', '_sidebar.md']
+      ignoreFiles: ['README.md', '_sidebar.md'],
+      docType: 'L2'
     },
     {
       gitURL: 'https://github.com/rdkcentral/Lightning-CLI',
       gitBranch: 'master',
       sourceDir: 'docs',
       targetDir: 'lightning-cli-reference',
-      ignoreFiles: ['README.md', '_sidebar.md', '.nojekyll', 'index.html']
+      ignoreFiles: ['README.md', '_sidebar.md', '.nojekyll', 'index.html'],
+      docType: 'L2'
     },
     {
       gitURL: 'https://github.com/rdkcentral/Lightning-SDK',
@@ -78,8 +83,31 @@ const config = {
       ignoreFiles: ['README.md', 'changelog.md', 'getting-started.md', '.nojekyll', 'index.html', 'package.json', '_sidebar.md'],
       typedocSourceDir: 'typedocs',
       typedocTargetDir: 'lightning-sdk',
-      typedocBuildRequired: false
+      typedocBuildRequired: false,
+      docType: 'L2'
     },
+    {
+      gitURL: 'https://github.com/lightning-js/renderer',
+      gitBranch: 'main',
+      targetDir: 'renderer',
+      sourceDir: 'docs',
+      typedocSourceDir: 'typedocs',
+      typedocTargetDir: 'renderer',
+      typedocBuildRequired: true,
+      docType: 'L3',
+      onlyTypedoc: true
+    },
+    {
+      gitURL: 'https://github.com/lightning-js/threadx',
+      gitBranch: 'main',
+      targetDir: 'threadx',
+      sourceDir: 'docs',
+      typedocSourceDir: 'typedocs',
+      typedocTargetDir: 'threadx',
+      typedocBuildRequired: true,
+      docType: 'L3',
+      onlyTypedoc: true
+    }
   ]
 };
 
@@ -105,11 +133,19 @@ async function getRemoteDocs(repo) {
 
     // generate typedoc (if they exist)
     if (repo.typedocSourceDir && repo.typedocTargetDir) {
+
       console.info(`Building TypeDocs for ${repo.gitURL}`);
       const tdSourcePath = path.join(clonedRepoPath, repo.typedocSourceDir);
       const tdTargetPath = path.join(config.typedocTargetBasePath, repo.typedocTargetDir);
+
+      // clean up existing temp repo & target dir & cloned repo path just in case the previous run failed
+      await fs.remove(tdTargetPath);
+
       shell.cd(clonedRepoPath);
-      const typedocCommand = `npm install ${repo.typedocBuildRequired ? '&& npm run build' : ''} && npm run typedoc`
+      const packageManager = repo.docType === 'L2' ? 'npm' : 'pnpm';
+      const typedocCommand = `${packageManager} install ${repo.typedocBuildRequired ? `&& ${packageManager} run build` : ''} && ${packageManager} run typedoc`
+      console.log(typedocCommand)
+
       const result = await exec(typedocCommand);
       if (result === 0) {
         await fs.move(tdSourcePath, tdTargetPath);
@@ -120,8 +156,10 @@ async function getRemoteDocs(repo) {
     }
 
     // move docs to target dir & delete all temp files
-    await fs.move(sourcePath, targetPath);
-    await Promise.all(repo.ignoreFiles.map(file => fs.remove(path.join(targetPath, file))));
+    if (!repo.onlyTypedoc) {
+      await fs.move(sourcePath, targetPath);
+      await Promise.all(repo.ignoreFiles.map(file => fs.remove(path.join(targetPath, file))));
+    }
     await fs.remove(clonedRepoPath);
 
     console.info('Finished getting docs from : ' + repo.gitURL);
@@ -143,40 +181,43 @@ async function generateSideMenu() {
 
   for (const repo of config.repos) {
 
-    console.info('Generating the menu for : ' + repo.targetDir);
+    if (repo.docType == 'L2') {
 
-    // read index.md file
-    const indexContent = await fs.readFile(path.join(config.targetBasePath, repo.targetDir, 'index.md'), 'utf8');
-    const tocMatches = indexContent.match(tocTextRegex);
+      console.info('L2-docs: Generating the menu for : ' + repo.targetDir);
 
-    if (tocMatches && tocMatches[1]) {
+      // read index.md file
+      const indexContent = await fs.readFile(path.join(config.targetBasePath, repo.targetDir, 'index.md'), 'utf8');
+      const tocMatches = indexContent.match(tocTextRegex);
 
-      // prepare toc content
-      const tocReplacements = [
-        { regex: /\r\n/g, replacement: '\n' }, // convert line endings to unix
-        { regex: /\t/, replacement: '  ' }, // replace tabs with 2 spaces
-        { regex: /^(.*)$/gm, replacement: '  $1' }, // add 2 spaces to the beginning of each line (list identation)
-        { regex: /\]\s*\((?!\/)/gm, replacement: '](/' + repo.targetDir + '/' }, // fix links coming from other index.md files, links that start with "/" are not replaced
-        { regex: /[ \t\r\n]+$/, replacement: ''}, // remove trailing spaces and line endings
-      ];
+      if (tocMatches && tocMatches[1]) {
 
-      let tocText = tocMatches[1];
-      for (const item of tocReplacements) {
-        tocText = tocText.replace(item.regex, item.replacement);
+        // prepare toc content
+        const tocReplacements = [
+          { regex: /\r\n/g, replacement: '\n' }, // convert line endings to unix
+          { regex: /\t/, replacement: '  ' }, // replace tabs with 2 spaces
+          { regex: /^(.*)$/gm, replacement: '  $1' }, // add 2 spaces to the beginning of each line (list identation)
+          { regex: /\]\s*\((?!\/)/gm, replacement: '](/' + repo.targetDir + '/' }, // fix links coming from other index.md files, links that start with "/" are not replaced
+          { regex: /[ \t\r\n]+$/, replacement: '' }, // remove trailing spaces and line endings
+        ];
+
+        let tocText = tocMatches[1];
+        for (const item of tocReplacements) {
+          tocText = tocText.replace(item.regex, item.replacement);
+        }
+
+        // add table of content text to the sidebar
+        const sideBarRepoLinkRegex = new RegExp('(^\\*.*\\(/' + repo.targetDir + '/.*?)$', 'm');
+        sideBarContent = sideBarContent.replace(sideBarRepoLinkRegex, '$1\n' + tocText);
+
+      } else {
+        console.error('L2-docs: TOC not found in ' + repo.targetDir + '/index.md');
+        throw ('L2-docs: TOCNotFound:' + repo.targetDir);
       }
-
-      // add table of content text to the sidebar
-      const sideBarRepoLinkRegex = new RegExp('(^\\*.*\\(/' + repo.targetDir + '/.*?)$', 'm');
-      sideBarContent = sideBarContent.replace(sideBarRepoLinkRegex, '$1\n' + tocText);
-
-    } else {
-      console.error('TOC not found in ' + repo.targetDir + '/index.md');
-      throw('TOCNotFound:' + repo.targetDir);
     }
   }
 
   // write sidebar file
-  console.info('Creating the new side bar menu file');
+  console.info('L2-docs: Creating the new side bar menu file');
   await fs.writeFile(config.sideBarFileTargetPath, sideBarContent);
 }
 
@@ -187,7 +228,7 @@ async function buildDocs() {
     await generateSideMenu();
   } catch (e) {
     console.error(e);
-    throw('documentBuildFailed');
+    throw ('documentBuildFailed');
   }
 }
 
@@ -198,72 +239,3 @@ buildDocs().then(() => {
   console.error(err);
   console.error('Building the documentation failed.');
 });
-
-const L3typedoc = {
-  tempDir: path.join(process.cwd(), '.temp-repos'),
-  targetBasePath: path.join(process.cwd(), 'public', 'api'),
-  repos: [
-      {
-          url: 'https://github.com/lightning-js/renderer',
-          branch: 'main',
-          targetDir: 'renderer',
-          sourceDir: 'typedocs',
-      },
-      {
-          url: 'https://github.com/lightning-js/threadx',
-          branch: 'main',
-          targetDir: 'threadx',
-          sourceDir: 'typedocs',
-      }
-  ]
-}
-
-async function getL3Typedocs(repo) {
-  const clonedRepoPath = path.join(L3typedoc
-  .tempDir, path.basename(repo.url, '.git'));
-  const targetPath = path.join(L3typedoc
-  .targetBasePath, repo.targetDir);
-
-  try {
-      await fs.remove(targetPath);
-      await fs.remove(clonedRepoPath);
-
-      await git
-          .clone(repo.url, clonedRepoPath)
-          .cwd(clonedRepoPath)
-          .checkout(repo.branch)
-
-      shell.cd(clonedRepoPath);
-
-      const tdSourcePath = path.join(clonedRepoPath, 'typedocs');
-      const tdTargetPath = path.join(targetPath);
-
-      const build = await exec('pnpm install && pnpm run build && pnpm run typedoc');
-      if (build === 0) {
-          console.info(`TypeDocs for ${repo.url} built successfully.`);
-      } else {
-          console.error(`TypeDocs for ${repo.url} failed with result:`, build);
-      }
-
-      await fs.move(tdSourcePath, tdTargetPath);
-      await fs.remove(clonedRepoPath);
-  } catch (e) {
-      console.error('An exception occurred while getting docs from :' + repo.url);
-      console.error(e);
-      throw ('repoCloningFailed');
-  }
-}
-
-async function buildL3Typedocs() {
-  try {
-    console.info('Starting building L3 typedocs');
-    await Promise.all(L3typedoc.repos.map(repo => getL3Typedocs(repo)));
-    await fs.remove(L3typedoc.tempDir);
-  }
-  catch(e) {
-      console.error('An exception occurred while getting docs from');
-      console.log(e)
-  }
-}
-
-buildL3Typedocs()
